+++ test/unit/test_paulis_with_placeholders.py
@@ -43,7 +43,7 @@ from pyquil.paulis import (
     sZ,
     ZERO,
 )
-from pyquil.quil import Program, address_qubits, get_default_qubit_mapping
+from pyquil.quil import Program, address_qubits
 from pyquil.quilatom import QubitPlaceholder
 
 
@@ -54,7 +54,6 @@ def test_simplify_terms():
     assert term.coefficient == -1.0
 
     term = PauliTerm("Z", q[0]) + PauliTerm("Z", q[0], 1.0)
-    print(str(term))
     assert str(term).startswith("(2+0j)*Zq")
 
 
@@ -247,14 +246,8 @@ def test_exponentiate_2():
     generator = PauliTerm("Z", q[0], 1.0) * PauliTerm("Z", q[1], 1.0)
     para_prog = exponential_map(generator)
     prog = para_prog(1)
-    print(prog)
     result_prog = Program().inst(CNOT(q[0], q[1])).inst(RZ(2.0, q[1])).inst(CNOT(q[0], q[1]))
-    print(result_prog)
 
-    mapping = get_default_qubit_mapping(prog)
-    print(mapping)
-    print("left:\n", address_qubits(prog).out())
-    print("right:\n", address_qubits(result_prog).out())
     assert address_qubits(prog).out() == address_qubits(result_prog).out()
 
 
@@ -658,7 +651,6 @@ def test_ordered():
     mapping = {x: i for i, x in enumerate(q)}
     term = sZ(q[3]) * sZ(q[2]) * sZ(q[1])
     prog = exponential_map(term)(0.5)
-    print(mapping)
     prog = address_qubits(prog, mapping)
     assert prog.out() == "CNOT 3 2\nCNOT 2 1\nRZ(1) 1\nCNOT 2 1\nCNOT 3 2\n"